requirements.txt
python-dotenv
aisuite>=0.4.0